+++ src/extract/rejection.rs
@@ -92,7 +92,7 @@ define_rejection! {
     #[status = UNPROCESSABLE_ENTITY]
     #[body = "Failed to deserialize the JSON body into the target type"]
     #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    /// Rejection type for [`Json`](super::Json).
+    /// Rejection type for [`Json`](crate::json::Json).
     ///
     /// This rejection is used if the request body is syntactically valid JSON but couldn't be
     /// deserialized into the target type.
@@ -103,7 +103,7 @@ define_rejection! {
     #[status = BAD_REQUEST]
     #[body = "Failed to parse the request body as JSON"]
     #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    /// Rejection type for [`Json`](super::Json).
+    /// Rejection type for [`Json`](crate::json::Json).
     ///
     /// This rejection is used if the request body didn't contain syntactically valid JSON.
     pub struct JsonSyntaxError(Error);
@@ -113,20 +113,77 @@ define_rejection! {
     #[status = UNSUPPORTED_MEDIA_TYPE]
     #[body = "Expected request with `Content-Type: application/json`"]
     #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
-    /// Rejection type for [`Json`](super::Json) used if the `Content-Type`
+    /// Rejection type for [`Json`](crate::json::Json) used if the `Content-Type`
     /// header is missing.
     pub struct MissingJsonContentType;
 }
 
 composite_rejection! {
-    /// Rejection used for [`Json`](super::Json).
+    /// Rejection used for [`Json`](crate::json::Json).
     ///
-    /// Contains one variant for each way the [`Json`](super::Json) extractor
+    /// Contains one variant for each way the [`Json`](crate::json::Json) extractor
     /// can fail.
     #[cfg_attr(docsrs, doc(cfg(feature = "json")))]
     pub enum JsonRejection {
         JsonDataError,
         JsonSyntaxError,
         MissingJsonContentType,
+    }
+}
+
+/// Rejection used for [`TypedHeader`](super::TypedHeader).
+#[derive(Debug)]
+pub struct TypedHeaderRejection {
+    pub(super) name: &'static http::header::HeaderName,
+    pub(super) reason: TypedHeaderRejectionReason,
+}
+
+impl TypedHeaderRejection {
+    /// Name of the header that caused the rejection
+    pub fn name(&self) -> &http::header::HeaderName {
+        self.name
+    }
+
+    /// Reason why the header extraction has failed
+    pub fn reason(&self) -> &TypedHeaderRejectionReason {
+        &self.reason
+    }
+}
+
+/// Additional information regarding a [`TypedHeaderRejection`]
+#[derive(Debug)]
+#[non_exhaustive]
+pub enum TypedHeaderRejectionReason {
+    /// The header was missing from the HTTP request
+    Missing,
+    /// An error occured when parsing the header from the HTTP request
+    Error(headers::Error),
+}
+
+impl IntoResponse for TypedHeaderRejection {
+    fn into_response(self) -> Response {
+        (http::StatusCode::BAD_REQUEST, self.to_string()).into_response()
+    }
+}
+
+impl std::fmt::Display for TypedHeaderRejection {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        match &self.reason {
+            TypedHeaderRejectionReason::Missing => {
+                write!(f, "Header of type `{}` was missing", self.name)
+            }
+            TypedHeaderRejectionReason::Error(err) => {
+                write!(f, "{} ({})", err, self.name)
+            }
+        }
+    }
+}
+
+impl std::error::Error for TypedHeaderRejection {
+    fn source(&self) -> Option<&(dyn std::error::Error + 'static)> {
+        match &self.reason {
+            TypedHeaderRejectionReason::Error(err) => Some(err),
+            TypedHeaderRejectionReason::Missing => None,
+        }
     }
 }
